feat: [vimrc] Read/Write to/from sys clipboard
To implement this feature new functions are added which should handle
the calls to system clipboards unobtrusively and silently handle any
errors.

NOTE: When read/write from/to system clipboard first it should
read/write from/to vim's default register i.e. '"' so that it eliminates
the need to write custom keymaps to handle various yank/paste cases such
as: paste before/after word, end of line, etc. This adds an extra to
step access the system clipboard but simplifies the vim usage when
yank/paste are required.