fix: [vimrc] Disable vim-markdown folding
As most markdown files are small and rarely need folding disabled this
option to speed up reading/writing these files. To enable we just need
to set this option 'let g:enable_spelunker_vim = 0' in command line and
refresh the editor with ':e'.

A custom function and keympa maybe needed to simplify this toggling this
option.