feat: Use vim as a pager for man command
The support for vim to be used as a pager for man is shipped with
default distribution. Add a custom shell function 'vman ()' to wrap
this feature in vim and use it for viewing man pages inside vim.
This is further enhanced with the support for command completions
of the new function 'vman ()' for zsh and bash.