fix: [shcmn-git] Use 'less' pager for grep
The issue with 'git-delta' as grep pager is that it changes the hyphens
'-' in some file names into ':' which breaks the grep functionality.
This is reported on github but not yet fixed so use git grep with less
and just add colors to file name and line number.

Remove commented 'delta' configurations and add a NOTE to document this
case.