feat: [vimrc] Enable vim spell check
Remove 'spelunker' spell check plugin and use vim built in spell check
as it provides insert mode spell suggestions and is a lot faster than
the plugin.

Due to enabling vim built in spell check setting dictionary is not
needed and dictionary completion will be provided from spell file. So
this is removed.

Update spell highlight colors to make it readable.