feat: [shell] Add common function source_plugins
This is done to make adding plugins to shells easy by calling the helper
function `source_plugins ()` instead of manually adding to the plugins
array and looping through it.

For this create a new shell common functions script to be sourced into
sh, bash or zsh and prevent duplicate code.